Use ticks_between_spectra SPEAD item
CBF have started providing this SPEAD item to indicate the timestamp
difference between consecutive spectra from the PFB. While this could be
calculated anyway (as 2 * n_chans) for the wideband correlator product,
there was not enough information to compute it for narrowband products
or for the beamformer with passband.

This commit uses this item in both the correlator and beamformer
capture. The --cbf-channels command-line argument to the beamformer is
deprecated and ignored; it will only be removed once the master
controller no longer tries to pass it.