Fix bug introduced by semantic merge conflict
Master and ar1.5-bf both changed the --cbf-spead option to use endpoint
list parser, and master then checked that it was a single endpoint and
dereferenced it. The latter made it into the merge, even though it is
incompatible with the ar1.5-bf code that expects it to be a list.