Increase bf_ingest window size
This controls the acceptable variation between arrival times of heaps
with the same timestamp. The previous value was working for ROACH but
broke down on SKARAB when the start of heap N+2 arrived from one engine
before the end of heap N from another.

This needs some testing to make sure the performance is acceptable,
since that buffer no longer fits in L3 cache.