Mostly revert changes
As per Bruce, katdal is only used for katdal.h5datav3.FLAG_NAMES and
FLAG_DESCRIPTIONS, and if you leave a pinned version then you don't
need to update this file every time the katdal or katpoint dependencies
change. The only change is now the base availability of dask and toolz.