Fix update of final stats
The last 10ms of statistics weren't being reflected into the counters.
That's not a major issue for production, but it was causing
non-deterministic unit test failures in test cases where the whole thing
took less than 10ms and so it thought that no data was received and
didn't write metadata into the file.

Also now ensures that the periodic update gets cancelled in the path
where the stream is terminated by a stop heap rather than a command.