Move argument parsing into the library code
This allows the unit test to use the argument parsing, instead of
providing a fully populated argparse.Namespace. The old approach was
painful as every new argument required an update to the test, even if it
had a sensible default, and in some cases the test values drifted away
from the argument defaults.