Prepare data writer for multiple heaps per slice
This follows on from preparing the receiver. Again, it is not actually
enabled, and may not work, but the code is in place. The data writer is
also converted to use the units framework.

The units framework is upgraded with a helper to make it easier to
define products between units.