Make bf_ingest test reliable
It was failing sporadically because despite using a reliable, in-order
transport, it used multiple instances of it and the order of delivery
was not determined between packets sent to different instances. It's
fixed by redirecting all the data that the receiver is actually expected
to receive into a single queue, so that it is ordered deterministically.

Closes SR-1762.