SPR1-2999 Enable cap_net_raw on capambel
This allows the product controller to run the script as
`capambel -c cap_net_raw+p -- ...` to use ibverbs.

The capambel tool is used to pass capabilities to the bf_ingest script
but it needs those capabilities itself. Since Docker 20.10.14 the
capabilities are not inherited anymore so set it explicitly.